refactor: share videogen rate limiter helpers
Move RateLimiterError and canister_request_key (now to_canister_request_key)
from generate.rs to rate_limiter.rs where they belong. Re-export
RateLimiterError from generate.rs via pub use for backward compatibility.
